Reenable `CONFIG=... ./deploy.sh cluster_auth`
The addition of `./deploy.sh set-project` broke this by requiring
the config to be explicitly included where necessary. This change
includes the config in all functions that need it, not just the
"official" entry points.

This pattern is used to configure kubectl to access the cluster of a
project given its config file. It can also be used to run an individual
deployment step when debugging deploy.sh.

Alternative considered:
Execute include_config when deploy.sh iff the config file exists.
However, this makes it harder to print a message like "ERROR: config.sh
doesn't exist" when the config is necessary but not present, or when
$CONFIG points to a non-existant file.

Change-Id: I33b11f5c94aa662caeb4e34922be3104e20d4fa6
GitOrigin-RevId: d2c282722f442f7d91bcd2b5f383ff23261007df