Sync finalizers along with status
This is required since ChartAssignment (CA) coordinate their deletion with
Helm by using a finalizer.
The CA controller on the robot accepts the CA and sets the finalizer
before install the Helm chart. If the finalizer isn't propagated to the
cloud, deleting the CA there would immedaitely cause it to disappear and
reporting of chart status's across the fleet would no longer be
accurate. In particular failing deletions on the robot could result in
undetected orphaned CAs.

Signed-off-by: Fabian Reinartz <freinartz@google.com>
Change-Id: Ia6d75f827bec4def342b6d2c7c5ce1184f6c63af
GitOrigin-RevId: 6eb0c54aaa55398e864309f1ff99b0935bdeef3c