Fix build for Bazel 0.23.0
Remove --incompatible_remove_native_* from .bazelrc. These are no-ops
for the build at this point, since we don't use the native rules. They
were added to bazelrc pre-0.20, to avoid accidental use of the native
rules, which is no longer a risk.

Also, fix the parameter ordering for core_app_chart, as *args and
**kwargs must appear last.

Closes #1.

Change-Id: Ib1844f7a30c05624774861d78387210b5f88795b
GitOrigin-RevId: 416ad9e22a1af9d0f37b3a1aebc86939c41c13b4