Handle status subresource and propagate pending deletion
If the status section of a CRD is declared an explicit subresource, a
dedicated call to UpdateStatus must be made on top of Update.

Syncing spec and metadata did not respect the deletion timestamp. This
causes a deletion deadlock if the deletion is not initiated downstream
but the upstream cluster waits for downstream deletion to complete.

Signed-off-by: Fabian Reinartz <freinartz@google.com>
Change-Id: I28ee626dc7d5c19a8a96549862a843997840874f
GitOrigin-RevId: a97b6c474514613a24d424c77cf4309670903892