Generate CRD YAML from proto.
This change introduces a Go binary that takes a protobuf message
with a few parameters and produces a CRD YAML file for Kubernetes. This
CRD YAML file can be passed to kubectl or put in a Helm chart for later
installation.

We put the proto descriptor into an annotation. The size limit for
annotations is 256k
(https://github.com/kubernetes/kubernetes/issues/15878), and a hello
world descriptor is 9k. We could probably reduce the descriptor size
with some pruning of unused messages.

Change-Id: Ia3c7708989075176ea305f8c81560fa0330c0409
GitOrigin-RevId: 7998b5e7466a15fa5cab1f06e90c31b327554994