Make deploy.sh set-project interactive
If the project_id is not given as parameter, ask the user to
enter the id of the GCP. When 'create' is called, but config.sh and
config.bzl are missing, calls 'set-project' first.

This is in preparation to use the same deploy.sh script to install
the binary, and to build and deploy from local sources. The goal
is to make the quickstart guide a 5min experience.

Change-Id: Id89ba0ca027e60476fe9de1980dc394e42b637fe
GitOrigin-RevId: 5b49bc3bfec46c9d6bb1b9f94fb7ecc519985f72