Read configuration from the CRD.
This removes the need to pass annotations as flags. Instead, the proto
descriptor is an annotation on the CRD, and the CRD is auto-generated at
build time and added to the cluster via Helm or kubectl.

I realize that this code badly needs unit tests by now and I promise
we'll follow up with them once it's integrated into data uploader.

b/124437335

Change-Id: Ic8670a917851c00db2a4073e0fb882899625feb6
GitOrigin-RevId: 9479d12b7707159cc0fd017398b6871399b76581