Request project number in setup-robot
This moves logic out of the shell script and save a command-line
parameter. This is a step towards removing all parameter-logic from
setup_robot.sh, which will decouple it from setup-robot and allow it to
pull the latest version of setup-robot instead of pinning a version.

Alternatives considered: I considered creating a new library to contain
`getProjectNumber`, but this goes against the Go proverb: a little
copying is better than a little dependency
(https://go-proverbs.github.io/).

b/124437977

Change-Id: If633a73ae24f12d12249f092cf565976349d962a
GitOrigin-RevId: a61aef9d203c07b91bb508471abd2117505c2e5b