Clean up deploy.sh
The script allowed to call all functions in the script which was not
document and prone to break when changes were made. Now only the real
commands are supported by the script. If we realize there is need to
support other commands, we can more commands.

Removed the hack that prevented build and push commands to be run when
deploying a binary. These commands are now bundeled in one function and
the branching is made explicit to avoid confusion. This also reduces
the number of bazel invocations to at most one per path.

Change-Id: Ifd62f14854031788eff4eaa32365d752d7afc922
GitOrigin-RevId: c70556fb4c93ec2b8933ff901316e9db4baae361