Handle version files in run-install.sh
gsutil does not support symlinks, therefore, version files are used to
link to the latest release. Later we can add version files for stable
and release candidates.

Specifying the bucket is not a main use case, so it won't be parsed from
the arguments anymore. This reduces complexity in the script.

Setting the config.sh in the run-install script is also supported
now.

Change-Id: Ib5db9c5a9f2568695b33a53619f741bdee7e05e2
GitOrigin-RevId: 656bfbbe588ae42412e10a4af7f29dea5fb9f1f7