Add a Layer 0 test
This test just asserts that `install_k8s_on_robot.sh` doesn't fail, so
we should consider adding assertions to check for a working cluster.

This requires adding a dependency on libretto, a convenient library that
provisions VMs, but is not actively maintained. However, it has worked
so far, and doesn't introduce so much complexity that we couldn't fork
it and fix any bugs we need to.

Change-Id: Ieb2fdd67aa0e961bf73dea6c8127911be1f84d2f
GitOrigin-RevId: a295877dad17673b0289719497f4529a779d23f3