Introduce a Method interface to cr-adapter.
This interface isolates the gRPC-handling code from the request-specific
code. It simplifies unit testing of the switch-heavy code in main.go.

resource_info and method are now one huge intertwined piece of code, but
their interface is easy to test. Let's cover with tests first, fix bugs,
and then detangle the code.

Change-Id: If71090c5ffc98361c06f38d5827c8ced2813c643
GitOrigin-RevId: 45ed175e80c41b23536c2f093d8bb39c642bb981