Support namespaces and cluster-scoped CRDs.
For all namespaced CRs, we add a "namespace" parameter to requests that
don't bring their own object. In the adapter, we honor the "namespace"
parameter in the request if the CR brings it, or default to "default".

To make the change easier, factor out a getStringField method that
descends into a dynamic proto along a path and returns the value of the
string field.

Change-Id: I21c20b5305fa2fe5fcd601b597c817c3c8dd9a0b
GitOrigin-RevId: ca62cb17823fad9f84bdb8c1096e4f383506e3a3