Add an example of a declarative API
This uses metacontroller to create a "charge-controller", that simulates
the charging of a robot. The controller logic is 50 lines of Python in
server.py, although we also need another ~100 lines of boilerplate.

This will be accompanied by a tutorial that explains how to install it,
how to use it with kubectl, and discusses the tradeoffs when using
metacontroller to wrap an external API.

Change-Id: Ica72b7940622b832dd2b70f0fe10104044470af1
GitOrigin-RevId: 0bebd11ae3f831995b61752820855ac890a5d8f3