Update to rules_go 0.17.1
Due to a change in how the dependencies work, we need to be careful to
load go_repository from deps.bzl, not def.bzl:

https://github.com/bazelbuild/rules_go/pull/1933

Additionally, Gazelle 0.17.0 seems to find the dependencies in
src/.gopath, so I've added a gazelle:ignore directive for them. At the
same time, I've added the path to .bazelignore, so that Bazel doesn't
try to build or run files in there.

This pulls in some Bazel compatibility fixes (#2).

Change-Id: Ic215ca30c0e6bb438e8fff134e439e4e4a032879
GitOrigin-RevId: 8c7708b9a97f5ab6a34db5aa36dc1be9b63d2fbd