Remove direct reference to genfiles
For historical reasons, Bazel has two directories for generated files,
although a future version of Bazel will remove the genfiles directory:
https://github.com/bazelbuild/bazel/issues/6761

Currently, rosmsg() copies all the source files (*.msg, *.srv) to the
output directory, so that the directory structure matches the needs of
ROS's codegen tools. It uses a hacky way of finding the paths to the
transitive dependencies within the output directory, which includes a
direct reference to genfiles. By using output_to_bindir=True, we can
avoid the direct reference.

Fixing the hack would involve rewriting the rosmsg() macro to use a
rule, which wouldn't be too hard but might break compatibility.

See #2.

Change-Id: Ibae3451e1329e397329089aa6796c5477789c788
GitOrigin-RevId: e38ef5d044984c858b7e1d21867347e1bb0d28a8