Stabilize and speed up the tests for resource_info.
Introduce an intermediate factory method BuildMethods that doesn't
depend on a Kubernetes fake so that we can test most logic without
invoking the Update function.

This allows us to split out the logic for Update into a dedicated test
that relies on gomega's "Eventually" tests (which polls for up to a
second). To simplify the code, drop the LogChannel/ErrorChannel logic
and instead just log directly to stdout again.

b/127252946

Change-Id: Ib5ecf39aa534ce08332dcd2ef5e9ee2c16928342
GitOrigin-RevId: be18d0de396fa03f16e616149ae6e05aba998480