Handle SIGTERM in example server code
SIGTERM doesn't terminate processes when they run as PID 1, which is the
case inside Docker containers.

This means that `kubectl delete pod` no longer takes ~30 seconds.
However, I'm not sure if this is worth the added complexity in the
examples. Additionally, doing a truly graceful shutdown seems to be more
complex:

https://medium.com/@fedor/shutting-down-grpc-services-gracefully-961a95b08f8
https://github.com/grpc/grpc/issues/10324#issuecomment-331581665

Finally, the C++ code has a race condition, as it will crash if SIGTERM
is received before the server is started, although this is fairly
unlikely and not harmful in this case.

Alternatives considered: using `kubectl delete --wait=false`

Change-Id: I04754181c56e5e5f86a64c89b1e16293889683ab
GitOrigin-RevId: 5ee5964d4605cdcd8780192b05c2f122c53f69d0