Don't require gRPC install in gRPC tutorial
This is a followup to an open comment in https://cloud-robotics-review.googlesource.com/c/cloud-robotics/+/26660

The tutorial required installing the gRPC library and the protoc tool
which was cumbersome if you just wanted to see things happening. Using
Docker to run the server and client locally avoids to install gRPC
locally. The downside is that the commands to start the client got a bit
more complicated.

Change-Id: Ia834a27c16712cca926bedeeabbdc03ffbb6d7ac
GitOrigin-RevId: e1e333baaf8b7a3417391d2e2077211797159d07