Add ros-master app to point to the ROS master
Currently all apps that use the ROS master need to define a value for
ROS_MASTER_URI, and all AppRollouts need to specify it. By defining this
for a single app, we reduce the boilerplate in apps that use ROS, since
they can refer directly to http://ros-master.app-ros-master:11311.

Alternatives considered:
- renaming the existing `ros` app, so we can use
  http://ros-master.app-ros:11311. However, we can always migrate to
  that later once the future of the current `ros` app is more certain.
- using Helm templating to write the correct endpoint into all apps that
  use ROS.

b/127601996

Change-Id: I36b764d70e813fbc1f688974ba785b16f15018e2
GitOrigin-RevId: c7563f9dc6ec407de36e6a59117de7ded9bfa118