Use cloud configs for binary installs
This allows using a config.sh and Terraform state in the cloud for
binary installs. The config is created interactively the first time the
installer runs and can be re-configured later. Users do not to keep any
state locally for binary installs and updates can be triggered from any
machine.

For source installs the same set-config script can be used to create the
config.sh and config.bzl.

Change-Id: I427cbd8efe0910cd9844d87610467f48caf88e42
GitOrigin-RevId: 938ad654e0b48743a44620305a6a8b2813b1f21d