Add mission crd app
The keypoint of the mission API is to have single declarative API for
controlling robots. The mission crd should be in a separate app so that
more than one mission controller can exist in the same cloud project.
Otherwise the first will install the CRD and the second will fail
because the CRD already exists. Even if the installing is successfull,
when removing the controllers, the first will delete the CRD and the
second will fail.

Change-Id: I8a8fc8bd780c814c2fbf29bfffaf066fd7773efd
GitOrigin-RevId: 91e6dbe392c3c2af971621f0a72d7d31bd3dccaa