Simplify use-synk switch for the app layer
Instead of enableding synk per AppRollout, the use-synk flag from
deploy.sh and setup-robot will now directly propagate through the app
layer and switch all apps to synk.

Due to upgrade issues with Helm, core CRDs are flagged to not be deleted
by Helm when the chart is removed. This requires people to do one
run of deploy.sh with USE_SYNK=false before switching to synk to avoid
issues.

Signed-off-by: Fabian Reinartz <freinartz@google.com>
Change-Id: I30458d55dc5455d60cad84467e6ffbc4cb249d3b
GitOrigin-RevId: a11c44f38f3d7494d93fa70cb708cfac98e176dc