Add a test that checks AppRollout status.
We run this before cehcking pod status. This will give us a better error
message (especially when an approllout failed). It will also defer the pod
checks due to the retried when waiting for the approllouts to settle.

Note: A first attempt used the generated go-client, but with that listing
approllouts always returned 0 and no error.

Tested by running:
bazel test \
  --test_tag_filters="external" \
  --test_env GCP_PROJECT_ID=XXXX \
  --test_env GCP_REGION=europe-west1 \
  --test_env GCP_ZONE=europe-west1-c \
  --test_env GOOGLE_APPLICATION_CREDENTIALS \
  --test_output=streamed \
  --spawn_strategy=standalone \
  //src/go/tests:go_default_test

Fixes #128978085

Change-Id: I5fa99a3454b396db5a816e693b173c67e7ec2c73
GitOrigin-RevId: 3114294de78b610f99bf10c0e1987a0650bcbcf1