Load variables from config.sh in setup-dev
This removes the dependency of setup-dev on the config.bzl and is part
of the effort to separate the build phase from the deploy/run phase.
setup-dev is no longer tied to one container registry after building.

Adds a library to read the config.sh from go programs for future development.

Change-Id: I72afca7ecfd2f425dcd027641c54275fd82470e7
GitOrigin-RevId: 4b6179b83bf39faea2f0737aab586f146560b5d4