Use headless service for ROS master on robot
I36b764d70 introduced the ros-master app, which aims to provide a DNS
name for the ROS master that works for both simulated and non-simulated
robots. However, this used an ExternalName service, which is inadequate
for non-simulated robots, as it can't use an IPv4 address for the
backend.

The k8s docs[1] recommend using a "headless service" for this case,
which was the approach used in I36b764d70. It is possible to make the
ros-master chart use a headless service or ExternalName service as
appropriate, albeit with some rather complicated templating.

Alternatives considered: rollback the ros-master app and use Helm values
to specify ROS_MASTER_URI for every app that uses ROS. If we have any
further trouble with the ros-master app, we should give up on it.

[1]: https://kubernetes.io/docs/concepts/services-networking/service/#externalname

Change-Id: I2498fce111be5cf3053d1ea7d3839e7a1028abdb
GitOrigin-RevId: 61844862f0ca0bf848977eb44c218138b6bef3a6