Make registry optional for app_chart
This is part of the effort to separate the build and push phases. In
particular, the build phase should not depend on the target registry.

While all app_charts are being migrated, setting the registry in
app_chart triggers a different build path for the push rules. This will
be removed after migration of all apps.

cloud-master and robot-master are migrated in this change as an example.

The scripts run_parallel, run_sequentially, and cache_gcr_credentials
propagate all incomming parameters to the invoked scripts. This allows
us to pass the container registry to the new push_all script.

Change-Id: Id2053640b0fbe6dddff120c571bba9e2918cb83f
GitOrigin-RevId: d49343fb1ba965cb44b5e9e3b84e65b9f5a66387