Pin build to Python version 2
With the upcoming change in Bazel 0.27 in how the Python runtime is obtained
(test with --incompatible_use_python_toolchains), the host configuration
by default will use Python 3. This causes the build to fail because
pip_import/pip_install for some reason still download the
Python-2-compatible version of PyYAML.

See #9.

Change-Id: I7bb5408263278114160fe0308490c5803914e03c
GitOrigin-RevId: 5ac81e3e0de9fa2523883fb9eeb3b1b17bce45ee