Update to latest grpc-java, rules_go and protobuf
This is required to use newer Bazel versions in CI.

Also grpc-gateway needs an update to cope with the use of virtual proto
imports in the newer protobuf release, and rules_docker needs to be
updated to cope with newer Bazel versions.

Updating rules_docker pulled in
https://github.com/GoogleContainerTools/distroless/pull/368, which
changes the default home directory in the containers to /root, which
requires a corresponding change in robot-master.yaml.

Change-Id: I958daa2356af795120e6b00a5f37b0c37d49d21d
GitOrigin-RevId: 61eedfd86e41abde069f35a33916c5c7daf47eff