Remove support to install charts with helm.
The chart-assignment controller supported both helm and synk as a backend
for a while. Since we switched to synk and in order to simplify the code
remove support for helm.

This also temporarily removes the release_test, since that is entierly
based on helm.. The test will be added back in a followup change.

Change-Id: Icaadeac0424ff1134d0e184f8ea5752bde65f8f8
GitOrigin-RevId: 872b2e96f79584f9b6c9d8b08042878bccf25322