Allow overriding the publicKeyRegistryID value.
If a developer wants to use minikube to create a test cluster on their
local machine, it is useful if the name of the kubectl context
(minikube) matches the name of the robot. However, if multiple such
developers run `setup-robot` with robot-name=minikube, they will
overwrite each others public keys in the cloud registry.

This allows them to override the ID for the cloud registry and prevent
such conflicts.

It is not easy to change the name of the kubectl context because
minikube's `none` driver doesn't support the `--profile` flag:
https://minikube.sigs.k8s.io/docs/reference/drivers/none/#other

Change-Id: I3cbc44316a7ee59015aed31215713797c616d475
GitOrigin-RevId: dc6e51f13707be42881fb16140cdc467ed37f487