Get the default grafana dashboards back.
When installing the prometheus operator using synk, we apply templates using
'helm template'. At k8s v14.0 things changes in kubernetes and prometheus
operator installs different verion of the dashboard depending on the k8s
server version. It uses the .Capabilities.KubeVersion.GitVersion value to
pick the right dashboard config. Unfortunately this version value is not
available to helm templat (only helm install).

Luckily someone else already run into it and the chart provides a
kubeTargetVersionOverride that we can use for this purpose.

Change-Id: If6134eee6dd03e1bbcf467cbefb53bc98e963a00
GitOrigin-RevId: 4605354c706d988976fca19e2f49be028808c3d6