Explicitly requeue checks for namespace deletion
If you delete & recreate a chartassignment that contains slow-to-delete
resources (eg containers that ignore SIGTERM) then the `app-<name>`
namespace will be asynchronously deleted[1]. The new chartassignment may
take ~one minute to reach the Updating state, even though the old
namespace only takes ~10s to be deleted.

This is becaues the chartassignment controller returns an error when
reconciling the new chart until the old namespace is gone.
controller-runtime rapidly retries this a certain number of times, then
starts a slow backoff that can take much longer than namespace deletion.
By explicitly requeuing the chartassignment, we control the frequency of
checking, ensuring that the new chartassignment is installed a few
seconds after the old namespace is deleted.

[1]: I think this is because synk uses background deletion to remove the
ResourceSet. I considered changing this but asynchronous deletion makes
it easy to build a new version of a chart while the previous version is
being deleted.

Change-Id: Ic3e9cfa3b22aacee31a4a5233c73296a93bcba00
GitOrigin-RevId: 477b4f84b27285eaff24dc5c3278c115e516d124