Allow modifying Jobs with synk
A Job's pod template is an immutable field, which causes problems when
Jobs exist in our charts. We can handle this case by detecting attempts
to mutate immutable Job fields, and handling these by deleting and
recreating the jobs.

This will have the downside that manual modifications to Jobs will be
lost (eg annotations, labels), but this seems fairly harmless, since
by far the most useful field in the Job is the immutable template.

Change-Id: Id402461bbdb4e0089c977c8588492dd56817ca80
GitOrigin-RevId: 3c9d495c7cdc1a48bb003b2db21d32cb62583fa9