Avoid reconcile loops because of status updates in the reconciler
The approllout/chartassignment controllers change the lastUpdateTime
on the conditions for every call to reconcile, regardless of whether the
spec has changed. The status updates trigger unnecessary calls to
Reconcile, and lead to log spam.

Instead, only update lastUpdateTime when the condition or message
changes, so the status is stable and Reconcile will not be called more
than necessary. This appears to be the approach used by the Deployment
manager.

Closes https://github.com/googlecloudrobotics/core/pull/42
https://github.com/googlecloudrobotics/core/pull/42

ORIGINAL_AUTHOR=oliver-goetz <o.goetz@sap.com>
GitOrigin-RevId: fede15eb5121ecd94fdde6e8ab6d6b719e20554b
Change-Id: I8be52659a7f6aaf70c33b3c73e9866eb77e9ebdd