Validate cluster names in ChartAssignments
If there is a bug in the AppRollout controller / cr-syncer, or if the
user manually creates a ChartAssignment, they could set a clusterName
that doesn't match the clusterName expected by the robot-master. In this
case, the robot-master will ignore the chartassignment, which is hard to
debug.

By amending the validator, we can create an error in this case, which
should help debug.

Change-Id: I4798266bdc66665a47b04777a20a70b33476c0e4
GitOrigin-RevId: 2422cb0eff81978a4b9468ad674f8c62709c1ceb