Use 60s timeout for requests to http-relay-server.
I observed a 7-minute relay outage, during which the relay-server
seemed to be operating as normal, but the clients were stuck in
remote.Get(). If that was caused by connection issues between the relay
client and server, it might be avoided by this change.

Go's default behavior is to not enforce a deadline, so it's not clear
how long the request could take to fail without setting a deadline.

Fixes b/157995266.

Change-Id: I2ca3627f76d1f5e9752c9b49c50b84dcc8f69a6b
GitOrigin-RevId: ed6ffae5215312ac66cfbaa557874879b431acfe