Don't retry with an invalid request ID.
If the http-relay-server restarts during a request, the relay client
might continue indefinitely trying to send response data that request,
even though the original client will have disconnected. The result is
lots of "Duplicate or invalid response id ..." messages in the log.
Avoid this by returning "400 Bad Request" in this case, and by treating
this as a permanent error.

Also, fix some uppercase error messages and reference to the request ID
as "response ID".

Fixes b/157995351.

Change-Id: I3497bfc33e5a79b3b107c49018ffa0b091f194cd
GitOrigin-RevId: 4248cc8d1219ecd410627d3e4dc5675875c63620