setup-robot: implement a check that prevents against accidentally overwriting the Robot.
Compare the hostname of the host of the cluster (passed in through an env var) with the "cloudrobotics.com/master-host" label. If they differ, ask user to confirm.
Because of https://github.com/kubernetes/kubernetes/issues/27264, sometimes the message is not printed if we log it too fast. I had to introduce a sleep in the (unlikely) new code path to combat this.

This also introduces a small change from creating a map[string]string to map[string]interface{} as I the k8s dynamic fake client paniced when trying to JSONify the former.

Fixes b/160772334

Change-Id: Ib1738cffa01921474cf8d7e6e18bfc5885a4054d
GitOrigin-RevId: a4f7d4fffb49160cd676068039e68b9bc604555d