Update path to the xyz_gomodules_jsonpatch repository as the gmodules.xyz/jsonpatch URL returns a 404.
I would have like to add an override to Gopkg.toml (similar to the entry for fsnotify) to make it clear that this is a "manual" override, but the dep tool then fails validation as it cannot look up the go-import metadata at the *original URL*. This is a bit counter intuitive but apparently WAI: https://github.com/golang/dep/issues/860. You can work around this by explicitly importing the .git path in your code, but in this case we don't own the code that uses this library.

Fixes b/168681321

Change-Id: I87f5bd1ae407d11d6defe7dc622792ec014d574c
GitOrigin-RevId: 2cb75f91de9bd04d7c803f5a82fc5c26f446edaa