Update protobuf package urls and version.
I tested the old urls with curl --head. This showed that the github repo
was moved. The new location is also part of the bazel mirror, but does not
have v3.11.2. It has v3.11.{3,4} instead and so this Cl bumps it to
v3.11.4.

The release notes only show minor fixes for c++, c#,php and ruby:
https://github.com/protocolbuffers/protobuf/releases/tag/v3.11.3
https://github.com/protocolbuffers/protobuf/releases/tag/v3.11.4

Change-Id: I05a4ecf310f77e223451804c4257f094e870a30e
GitOrigin-RevId: b471d5c4ebe226a65282a864a0e2a6146a8de2d7