Treat "namespace being deleted" as transient
This should reduce the severity of a race condition between chart
deletion (that leads to async namespace deletion) and chart creation
(that will recreate the namespace when the previous one is gone).

We can't rely on the status code as the apiserver returns a 403
Forbidden, which is normally not transient, so we have to check the
message text. Later versions of Kubernetes include a status cause, but
we would like this to work with v1.16.

Change-Id: I3c2267b856bc050dd4f101383b754691007c9bea
GitOrigin-RevId: 5bb91d70fb6838a2a67a7f91e6c064b88a09fdb4