synk: Improve CRD support
This adds support for the newer v1 CRDs that specify multiple versions
instead of the deprecated single version field. It also pulls the
invalidation of the discovery cache out of a for loop, which improves
the runtime on charts containing many CRDs from ~20s to ~1s.

Finally, it improves some error messages that I encountered when working
on this.

Change-Id: I7e1946a89975e1c5c18f301d3be5ba5afa98b363
GitOrigin-RevId: 49a6446853a10c1868fd97d7d4a7549c02931173