Back out overwrite-robot check added in a4f7d4fffb49160cd676068039e68b9bc604555d.
Turns out this is probably causing more harm than good: it has a tendency to trigger false positives when whatever is past in the master-host label doesn't match HOST_HOSTNAME i.e. $(hostname).
I am still keeping the logic to merge the labels when updating a robot and the corresponding test, as there are situations when you want to add a label after initial creation.

FIXED b/170370848

Change-Id: I70ce045679fc14bcbb84dffcba223acdbe117337
GitOrigin-RevId: c681f7162e63735dacdd991dfa6033042948f373