Apply crc-version tag during setup-robot.
Fetch the crc-version tag along with the install script and pass it to the
setup-robot image. There add it as annotation to the robot cr.
This will let us compare the annotation on existing robot instances to the
crc-version tag in the cloud bucket to see which instances are out-of-date.

Change-Id: I21e502ca35b009ad45b43fdb641d7f5a8884a89e
GitOrigin-RevId: 850185001aa36781a0ff2e97ac2201a7e9cc8105