Add RBAC policy for cr-syncer
This will allow us to reduce the IAM permissions granted to
robot-service@, preventing a compromised robot from running workloads in
the GKE cluster or on other robots.

I've enabled the /status subresource on Robots, preventing one robot
from changing the type or labels of other robots, which could be used,
for example, to start a vulnerable pod on robots that wouldn't run it by
default.

Change-Id: I5d9c5396d6654656e0443a0f9090b2ddd42da447
GitOrigin-RevId: 161823db2d434b15b111c217865f95887d1e823d