Set flush_interval to 1s for fluentd concat plugin.
This should somewhat mitigate issues when unrelated log lines are concatenated because lines that use the google style log prefix (that our regexp looks for) are interleaved with lines that don't.
The hypothesis is that most log entries that are actually supposed to be multiline will be output with <1s between lines.

Bug: b/178191140
Change-Id: I3129b94e5d5b8b686b788534821cc62bb074e17e
GitOrigin-RevId: f8eaa66ab62716c31820ffd9ac0af69fd1d8430f