Update to ingress-nginx v0.44.0
The newer version will gracefully close the connection to the cr-syncer
when the config reloads, so this goes some way towards fixing #64
(although the cr-syncer should handled dropped connections better).

The Deployment is based on upstream's deploy.yaml, but without RBAC
policies, the namespace and the service account, to avoid changing too
much. The newer Deployment drops privileges and has more graceful
shutdown behavior.
https://github.com/kubernetes/ingress-nginx/blob/master/deploy/static/provider/cloud/deploy.yaml

This requires adjusting the ingress due to the following breaking
changes (which, incidentally, despite the "breaking changes" category in
the changelog, were not listed as breaking...):

- the secure-backends annotation has been removed
- path rewriting now uses regex groups rather than prefixes

While we're here, we can stop using the deprecated extensions/v1beta1
Ingress.

Change-Id: I24c266f21e489a5c9079e9f884699c84d6be6a0e
GitOrigin-RevId: 564dd8042320d6becf40082911415eee849de7ee