Upload syslog
This will let us audit SSH access as well as debugging
network/cron issues that are visible in syslog.

We upload the non-standard /var/log/syslog.fluentd because the standard
files are all in RFC-3164 format, which doesn't specify a timezone.
Since fluentd runs in a container, it can see a different timezone to
the host and incorrectly interpret the RFC-3164 log timestamps.

Fixes b/182541978.

Change-Id: Ice43387159e26d991ff85f21e0138069b287ec11
GitOrigin-RevId: 45d0fef249e84af886c73cbd4c982830dbf86be2