Wrap kubectl run in a TTY
`kubectl run -it --rm` only works properly in a TTY, otherwise the logs
are not printed properly, and it will fail if the command takes longer
than 60s. I considered trying to recreate the effect of `-it --rm` with
kubectl wait/logs/delete, but this seems easier.

Also, this adds a "Setup complete" message to setup robot to make
success more obvious.

Change-Id: Id73d2938df5ab0fbf5a26e0728558407fb289b64
GitOrigin-RevId: 5fe0e6418b7e30d7c68347099400641bce7e4ab7