Allow charts to install into kube-system.
Several helm charts do this to customize or conenct to existing installs (e.g.
prometheus-operator and cert-manager). The strictness of the check currently
required us to disable some of the features in those charts or adding
workarounds.

We only add an exception for kube-system sicne ehre we can be pretty sure
that it exists.

Change-Id: I285330fc05cc55bd5ea4399694427b287bb18503
GitOrigin-RevId: 32009569bf626725b59c81c7b8cfb142c1809add