Fix collection of the kublet metrics.
We had disabled the service, claiming that it is already provided by k8s,
which is neither the case for minimuke, GKE or any of the other setups we
use.

We also replaced the ServiceMonitor for the kubelet with a custom one
claming that the default one does not work for GKE. Despite this, checking
the scrape targets in the prometheus web-ui shows 0/0 targets for the
kubelet.

Backing out both changes makes the targets show up and data coming in.
This can be verified also in Grafana using the "Kubernetes / Compute
Resources / Cluster" dashboard.

See b/190160361

Change-Id: If1378e3e2f4dd4530ce2a1297a59d8d34ce4a8ed
GitOrigin-RevId: 30586786705f21454980a7cdbd4835d8b41af512