Add a num_pending_requests flag to the relay client to allow enqueuing parallel requests.
When using the relay for a web-frontend previously only one request would be
sent and once the server starts using it by replying, we enqueu another
request.

This change adds a flag to the relay client that is used to enqueue multiple
requests on the server side. This should reduce the latency when doing many
small requests.

The server is already using a channel and block on available clients and
hence needs no change.

Change-Id: I8dcf40348b3efe553f72491b442d16e9ae3cf5b6
GitOrigin-RevId: 68aba6f1cf707fd7ab5549f0e3a1368761be3c59