Enable workload identity on the cloud-robotics cluster.
This will allow us to use more fine-grained permissions and is also a
pre-requisite for e.g. mcs (multi-cluster-services). In order to use
it we need to ensure the cluster are using VPC natie networking - a
feature that will become default for new cluster "soon" anyway. Here
we let GCP pick suitable networks, which is okay since our clusters
are all relative small.

See b/175282543

Change-Id: I53420e3b04b44e69742161db809259e25fec0d01
GitOrigin-RevId: 192f004cd9097678a2dd6c0bc7a0bac37706bdb6