Set a 60s timeout on cr-syncer health checks
This is longer than the default request timeout of 30s, so I hope it
will fix #67. Unfortunately, we're building against an older client-go
that doesn't support request contexts on the dynamic client, but I don't
think the long timeout will cause a problem since we're intentionally
avoiding a failing health check for network issues.

Change-Id: I1989a8a73c49ab068a5d848defc6b1388ccfaf60
GitOrigin-RevId: baca4440b706939155a2535d2607dfb225726bf6