Improve resilience of cr-syncer regarding network errors
@drigz, this is the follow up to issue https://github.com/googlecloudrobotics/core/issues/67

It changes the limiter of the queue which is source for changes in cloud cluster to `ItemFastSlowRateLimiter` that there is no exponential backoff. Exponential backoff could be the reason for long recovery times on unstable network connectivity.
`ItemFastSlowRateLimiter` is configured to do 5 fast retries with 500ms backoff time, followed by a series of slow retries with 5 seconds backoff time.

The limiter of the queue for updating CRs in der robot cluster remains `DefaultControllerRateLimiter` because we don't expect connectivity issues there.

Closes https://github.com/googlecloudrobotics/core/pull/68
https://github.com/googlecloudrobotics/core/pull/68

ORIGINAL_AUTHOR=oliver-goetz <o.goetz@sap.com>
GitOrigin-RevId: 6d6a5a27f15814343c4bf058eb19a33099008c35
Change-Id: I3da1c63090386767f20b60895af89b3af706de33