Remove ${project}-tools bucket.
This was used in the past for some setup tools. In newly setup projects it
is empty. I also don't see any of the tests referencing it. Since we have
set 'force_delte' on teh bucket, we can just remove it and it will be
deleted with all the content.

Change-Id: I9b6449543eb18f73f684b08c8d0d4ba08eb7147a
GitOrigin-RevId: 84381de95afb52a29a5b46da6f9974561daac3b7