Update rules_go and gazelle
We had to move the container rules to the end of the workspace, since
they otherwise pulled older versions of go libraries (noteably x_net).
Also a few dependencies had to be updated to fix build issues (incl.
bazel 3.7 -> 4.0).

Finally We need a workaround for gazelle - this used to be run via
bazel, but for now one needs to go get it and run it directly.

Change-Id: I0125128413297c07d6a2a19da4f282ac08ebb718
GitOrigin-RevId: adfccede7b45927cd8e7d836f68772931ea938c6