setup-robot: accept a pod CIDR.
In this context, "CIDR" means "IP address range". In the Kubernetes
ecosystem, "CIDR", "IP range", and "subnet" seem to be used
interchangeably, so I picked the one I've heard most often.

The metadata-server will only accept requests from this CIDR, so we need
to plumb it through from setup-robot to the metadata-server.

Change-Id: I53821804e81e4c7cfd5d376a488589369b670112
GitOrigin-RevId: f2b1072e844bbd9cf3d8ff57fd7fd29b2d81abf1