Add timeout option for cr-syncer
We are still facing some connection issues using istio ingress. The error message suggests that the reason for these issues could be closed connections between ingress and kubernetes api server. Even though the message tells us "NO_ERROR" the connection is closed afterwards.
```
W1013 22:20:07.214926       1 reflector.go:441] go/cmd/cr-syncer/syncer.go:259: watch of *unstructured.Unstructured ended with: an error on the server ("unable to decode an event from the watch stream: stream error: stream ID 3605; NO_ERROR") has prevented the request from succeeding
```

I found a good option to deal with those kinds of issues by introducing a timeout for the watch call.
It is implemented at two places
1. The options of List & Watch function
2. The remote client

When implementing the timeout just at the watch function. it is included into an URL parameter of the watch call. With this parameter set, the API server is closing the connection, but the timeout is not enforced on client side. Thus, if the connection breaks before the timeout cr-syncer does not recognize it.
By default the timeout url parameter seems to set to a random value between 4 and 10 minutes (estimated).

Timeout is disabled by default, so cr-syncer is not able to recognize an eventual timeout by himself.
When implementing it in the remote client only, there are warnings on each reconnect
```
W1015 15:52:02.629074       1 transport.go:288] Unable to cancel request for *main.ctxRoundTripper
```

That's why I implemented it at both places.
There is a new parameter `--timeout` where you can set the timeout in order to optimize the settings. By default it is set to 300 seconds. This should not stress the api server too much.
I set the timeout in the client to `timeout +5` to give it some time for a graceful closing.

Additionally, I removed the `configureTransport` function which is now the default setting at http client in apimachinery.
Btw. this feature is not detecting the lost connection in our case. Probably the istio ingress is answering the http ping requests even when its connection to the backend is broken.

Closes https://github.com/googlecloudrobotics/core/pull/84
https://github.com/googlecloudrobotics/core/pull/84

ORIGINAL_AUTHOR=oliver-goetz <o.goetz@sap.com>
GitOrigin-RevId: fe9187dbe9f246f93ff8210775a3b65b23d2edbf
Change-Id: Ib439147e6b626a7993e11145461fe5359d96c885