Add some checks to track down where we endup with invalis resource sets.
We've seen cases that fail with this error:
```
create resources object ".v1": ResourceSet.apps.cloudrobotics.com ".v1" is invalid:
metadata.name: Invalid value: ".v1":
a DNS-1123 subdomain must consist of lower case alphanumeric characters, '-' or '.',
and must start and end with an alphanumeric character (e.g. 'example.com', regex
used for validation is '[a-z0-9]([-a-z0-9]*[a-z0-9])?(\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)*')
```

Sprinkle some fatal logs to get backtraces.

Change-Id: Ic8de0d96b3f54ade17452d538ced78734522b980
GitOrigin-RevId: 42d9f49344ec37afa22b72368107c54561aaaebd