Enable RBAC in the cloud cluster
This PR adds cluster roles and for app-rollout and
chart-assignment-controllers and starts them with dedicated service
accounts. **Additionally the permissive-binding is removed during
deployment.**

The cluster roles for app-rollout-controller are pretty easy and just
include additional permissions for `robots`, `apps`, `approllouts` and
`chartassignments` CRDs.

Chart-assignment-controller has now permissions for `chartassignments`
and `resourcesets` CRDs and for namespaces. For convenience the
`cluster-admin` role is assigned to its service account too, because we
do not really have a concept yet, which Kubernetes objects we would like
to be managed by chart-assignment-controller.

However, this change allows to investigate roles and permissions of
other apps now, which was somehow pointless while a permissive-binding
was existing. Thus, this PR might lead to "permission denied" in other
apps.

Closes https://github.com/googlecloudrobotics/core/pull/85
https://github.com/googlecloudrobotics/core/pull/85

ORIGINAL_AUTHOR=oliver-goetz <o.goetz@sap.com>
GitOrigin-RevId: 3657a2c5086705266be56ed03b19cdbcf377f9a8
Change-Id: I1a09ad6f64c317e0d9029dbfe317c8543ac6d455

Change-Id: I1dc58f2e447299699c1cf647031c2c946720387e
GitOrigin-RevId: cd8b56397293afaeee3ca3e17cc6197f039bb3eb